# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for AvitoProj Application",
        "title": "Avito Proj API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/api/sections": {
            "post": {
                "description": "create new section",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sections"
                ],
                "summary": "Create new section",
                "operationId": "create-section",
                "parameters": [
                    {
                        "description": "section (slug) name",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.SectionRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: Section created",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "400": {
                        "description": "message: Failed",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete section",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sections"
                ],
                "summary": "Delete section",
                "operationId": "delete-section",
                "parameters": [
                    {
                        "description": "section (slug) name",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.SectionRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: Section deleted",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "400": {
                        "description": "message: Failed",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "/api/sections/users/:id": {
            "get": {
                "description": "get user sections",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sections-users"
                ],
                "summary": "Get user sections",
                "operationId": "get-user-sections",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "put": {
                "description": "add or delete user in sections",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sections-users"
                ],
                "summary": "Add user in section",
                "operationId": "add-user",
                "parameters": [
                    {
                        "description": "list of sections to add or delete",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.AddDeleteSections"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: Success",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "400": {
                        "description": "message: Failed",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "/api/users": {
            "post": {
                "description": "create new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create new user",
                "operationId": "create-user",
                "parameters": [
                    {
                        "description": "user's name",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UserRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: User created",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "400": {
                        "description": "message: Failed",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "/api/users/:id/history": {
            "post": {
                "description": "get user history",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user history",
                "operationId": "get-user-history",
                "parameters": [
                    {
                        "description": "date ('YY-MM')",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.HistoryRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/avitoproj.UserHistory"
                            }
                        }
                    },
                    "400": {
                        "description": "message: Failed",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "avitoproj.UserHistory": {
            "type": "object",
            "properties": {
                "operationDate": {
                    "type": "string"
                },
                "operationType": {
                    "type": "string"
                },
                "sectionId": {
                    "type": "integer"
                },
                "sectionName": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "handler.AddDeleteSections": {
            "type": "object",
            "properties": {
                "sectionsAdd": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "sectionsDelete": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "handler.HistoryRequestBody": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                }
            }
        },
        "handler.SectionRequestBody": {
            "type": "object",
            "properties": {
                "slug": {
                    "type": "string"
                }
            }
        },
        "handler.UserRequestBody": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        }
    }
}